visualizers can be changed via the ui

+++ src/core/visualizer/ISimulationVisualizer.hpp
@@ -22,8 +22,8 @@ namespace FluidSolver {
 		};
 
 		struct Size {
-			size_t width = 1920;
-			size_t height = 1080;
+			uint32_t width = 1920;
+			uint32_t height = 1080;
 		};
